Refactor snapshotting
Should be way faster now.

The way things used to be:
Snapshotting would iterate over all files with a TreeWalk. For each
file, it would iterate over all the commits in the repo to find which
one contained the most recent change to that file. This was slow mainly
because RevWalk instances can’t be reused (see:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=439856)

The way things are now (as of this commit):
List all paths in the given commit’s tree and build a hash of paths to
nil values. Use a single RevWalk to iterate over all commits. Use a
single instance of TreeWalk to find the files changed in that commit
(reset it on each iteration). Gradually replace nil values in the
snapshot hash. When all commits have been examined, the hash shouldn’t
contain any nil values. Note: TreeWalk instances can be reused.