Introduce skeleton HeadSnapshotFactory
Should grab a “snapshot” of the current HEADs and ignore merges. In
other words, if the current tip of my_branch is a merge commit,
HeadSnapshot factory should search for the next child commit that is
NOT a merge and return that instead. This is consistent with how the
rest of Rosette works, i.e. always ignoring merge commits.